sample_db_pg/requirements.txt
psycopg2-binary==2.9.11
faker==37.12.0
